// font-end/src/types/customer/group.ts
export enum GroupRole {
  Member = "Member",
  Moderator = "Moderator",
  Admin = "Admin",
}